requirements.txt
# FastAPI Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25

# HTTP Client
requests==2.31.0

# Scheduled Tasks
apscheduler==3.10.4

# CORS is built into FastAPI

# DateTime Utilities
python-dateutil==2.8.2
